pyproject.toml
[project]
name = "tootlogger"
version = "0.4.4"
description = "Log your Mastodon toots to DayOne"
readme = "README.md"
requires-python = ">=3.11,<4"
license = "BSD-3-Clause"
authors = [{ name = "Amelia Aronsohn", email = "squirrel@wearing.black" }]
keywords = [
  "mastodon",
  "social media",
]
classifiers = [
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "Natural Language :: English",
  "Topic :: Utilities",
]
dependencies = [
  "Jinja2>=3.1.2",
  "html2text>=2020.1.16",
  "mastodon.py>=1.6",
  "tomli-w>=1.0",
]

[project.urls]
Repository = "https://onlyhavecans.works/tools/tootlogger"

[project.scripts]
tootlogger = "tootlogger.cli:main"

[dependency-groups]
dev = [
  "basedpyright>=1.22",
  "pytest>=8.0",
  "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"

[tool.basedpyright]
typeCheckingMode = "standard"
pythonVersion = "3.13"
